+++ tests/tasks/test_re.py
@@ -1,3 +1,7 @@
+import os
+
+import pytest
+
 from rex.tasks.relation_extraction import (
     MCBagRelationClassificationTask,
     MCMLSentRelationClassificationTask,
@@ -5,6 +9,7 @@ from rex.tasks.relation_extraction import (
 from rex.utils.config import ConfigParser
 
 
+@pytest.mark.skipif(not os.path.exists("data/NYT10/formatted"), reason="data not found")
 def test_bag_re_task():
     cmd_args = ["-dc", "conf/re/bag_nyt10.yaml"]
     config = ConfigParser.parse_cmd(cmd_args=cmd_args)
@@ -14,6 +19,7 @@ def test_bag_re_task():
     task.train()
 
 
+@pytest.mark.skipif(not os.path.exists("data/IPRE/formatted"), reason="data not found")
 def test_sent_re_task():
     cmd_args = ["-dc", "conf/re/sent_ipre.yaml"]
     config = ConfigParser.parse_cmd(cmd_args=cmd_args)